The following changes were made in the commit: 0) Reformatting and Optimized Imports by CLion 1) Changed all methods to use float instead of double when working with coordinates since OpenGL supports floats 2) Changed the range of random numbers in testing to around +-1E+-39 3) Added Renderable Object class which can take in a rigidBody and gives its coordinates and indices to work with OpenGL We will write a test file for Renderable Objects in the next commit

+++ include/Kernel/Polygon.hpp
@@ -4,7 +4,7 @@
 #include <eigen3/Eigen/Dense>
 #include <vector>
 
-using Eigen::Vector2d;
+using Eigen::Vector2f;
 
 namespace Isaac::Kernel
 {
@@ -14,38 +14,38 @@ namespace Isaac::Kernel
     
     Polygon() noexcept;
 
-    Polygon(const int& vertexCount_, const std::vector<Vector2d>& vertices_, const Vector2d& position_, const Vector2d& velocity_, const Vector2d& acceleration_) noexcept;
+    Polygon(const int& vertexCount_, const std::vector<Vector2f>& vertices_, const Vector2f& position_, const Vector2f& velocity_, const Vector2f& acceleration_) noexcept;
 
     Polygon(const Polygon& polygon_) noexcept;
 
-    Vector2d getPosition() const noexcept;
+    Vector2f getPosition() const noexcept;
 
-    void setPosition(const Vector2d& newPosition) noexcept;
+    void setPosition(const Vector2f& newPosition) noexcept;
 
-    Vector2d getVelocity() const noexcept;
+    Vector2f getVelocity() const noexcept;
 
-    void setVelocity(const Vector2d& newVelocity) noexcept;
+    void setVelocity(const Vector2f& newVelocity) noexcept;
 
-    Vector2d getAcceleration() const noexcept;
+    Vector2f getAcceleration() const noexcept;
 
-    void setAcceleration(const Vector2d& newAcceleration) noexcept;
+    void setAcceleration(const Vector2f& newAcceleration) noexcept;
 
     int getVertexCount() const noexcept;
 
-    Vector2d getVertex(const int& n) const noexcept;
+    Vector2f getVertex(const int& n) const noexcept;
 
-    void setVertex(const int& n, const Vector2d& newVertex) noexcept;
+    void setVertex(const int& n, const Vector2f& newVertex) noexcept;
 
-    std::vector<Vector2d> getVertices() const noexcept;
+    std::vector<Vector2f> getVertices() const noexcept;
 
-    void setVertices(const std::vector<Vector2d>& newVertices) noexcept;
+    void setVertices(const std::vector<Vector2f>& newVertices) noexcept;
 
   private:
-    Vector2d position;
-    Vector2d velocity;
-    Vector2d acceleration;
+    Vector2f position;
+    Vector2f velocity;
+    Vector2f acceleration;
     int vertexCount;
-    std::vector<Vector2d> vertices;
+    std::vector<Vector2f> vertices;
   };
   
 } // namespace Isaac::Kernel
