+++ include/Kernel/RigidBody2D.hpp
@@ -15,27 +15,27 @@ namespace Isaac::Kernel {
 
         Polygon getPolygon() const noexcept;
 
-        Vector2d getPosition() const noexcept;
+        Vector2f getPosition() const noexcept;
 
-        void setPosition(const Vector2d& newPosition) noexcept;
+        void setPosition(const Vector2f& newPosition) noexcept;
 
-        Vector2d getVelocity() const noexcept;
+        Vector2f getVelocity() const noexcept;
 
-        void setVelocity(const Vector2d& newVelocity) noexcept;
+        void setVelocity(const Vector2f& newVelocity) noexcept;
 
-        Vector2d getAcceleration() const noexcept;
+        Vector2f getAcceleration() const noexcept;
 
-        void setAcceleration(const Vector2d& newAcceleration) noexcept;
+        void setAcceleration(const Vector2f& newAcceleration) noexcept;
 
         int getVertexCount() const noexcept;
 
-        Vector2d getVertex(const int& n) const noexcept;
+        Vector2f getVertex(const int& n) const noexcept;
 
-        void setVertex(const int& n, const Vector2d& newVertex) noexcept;
+        void setVertex(const int& n, const Vector2f& newVertex) noexcept;
 
-        std::vector<Vector2d> getVertices() const noexcept;
+        std::vector<Vector2f> getVertices() const noexcept;
 
-        void setVertices(const std::vector<Vector2d>& newVertices) noexcept;
+        void setVertices(const std::vector<Vector2f>& newVertices) noexcept;
 
         double getMass() const noexcept;
